Refactor `Component` and `SingletonComponent` derive macros
